Correctly handle the cancelled case.
When the user taps "Show Answer", the question is considered
"cancelled" - even if they get it right, it doesn't count, and their
correct attempts count will be reset.
The question should not advance to make it clear that this has
happened and also to give an extra change to look at the answers.