Mark 3 questions as multiple answers
There are three questions that aren't bulleted for answers,
so just saying them in the given order would work, but there's
no reason to be that limited.  Mark them as multiple answers.

Upcoming work to allow multiple answers in a single sentence will
allow the old-style answers to work as well.