Reverse matching algorithm
Instead of making a new sublist and passing it forward each time,
expect it to come in reversed and pop the last item off the list.

This causes the original list to get modified but is a way of returning
the remainder to the calling function.

This is part of a larger change to the matcher that's coming.

Tested:
Modified tests, they all pass.