Implement multiquestionwalker
This is an interable that will return the correct set of answers as it
walks through a string.

This is useful for multiple-result questions.  Two cabinet positions:

"Secretary of state and secretary of agriculture"

They can come in any order, and there's a number to choose from.
This walks the tokens until it's found a good match and returns those.